+++ CHANGELOG.md
@@ -7,6 +7,12 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 -->
 
+## [0.6.1] - 2025-04-03
+
+### Bug fixes
+
+* Fixed a missing dependency on the `requests` package.
+
 ## [0.6.0] - 2025-04-01
 
 ### New features